chore: renamed package models to model
